Artist routes ready

+++ internal/services/artist_service.go
@@ -3,12 +3,15 @@ package services
 import (
 	"chords_app/internal/models"
 	"chords_app/internal/repositories"
+
+	"errors"
 )
 
 type ArtistService interface {
 	CreateArtist(name, description, imageUrl string) (*models.Artist, error)
 	UpdateArtist(artistId uint, name, description, imageUrl string) (*models.Artist, error)
 	DeleteArtist(artistId uint) error
+	GetArtists() (*[]models.Artist, error)
 	GetArtistInformation(artistId uint) (*models.Artist, *[]models.Song, error)
 }
 
@@ -16,7 +19,7 @@ type artistService struct {
 	repo repositories.ArtistRepository
 }
 
-func NewArtistRepository(repo repositories.ArtistRepository) ArtistService {
+func NewArtistService(repo repositories.ArtistRepository) ArtistService {
 	return &artistService{repo}
 }
 
@@ -43,10 +46,19 @@ func (s *artistService) UpdateArtist(artistId uint, name, description, imageUrl 
 	if err != nil {
 		return nil, err
 	}
+	if artist == nil {
+		return nil, errors.New("artist not found")
+	}
 
-	artist.Name = name
-	artist.Description = description
-	artist.ImageUrl = imageUrl
+	if name != "" {
+		artist.Name = name
+	}
+	if description != "" {
+		artist.Description = description
+	}
+	if imageUrl != "" {
+		artist.ImageUrl = imageUrl
+	}
 
 	err = s.repo.UpdateArtist(artist)
 	if err != nil {
@@ -61,6 +73,9 @@ func (s *artistService) DeleteArtist(artistId uint) error {
 	if err != nil {
 		return err
 	}
+	if artist == nil {
+		return errors.New("artist not found")
+	}
 
 	err = s.repo.DeleteArtist(artist)
 	if err != nil {
@@ -72,13 +87,16 @@ func (s *artistService) DeleteArtist(artistId uint) error {
 
 func (s *artistService) GetArtistInformation(artistId uint) (*models.Artist, *[]models.Song, error) {
 	artist, err := s.repo.GetArtistById(artistId)
-	var empty_songs *[]models.Song
-
 	if err != nil {
-		return nil, empty_songs, err
+		return nil, nil, err
+	}
+	if artist == nil {
+		return nil, nil, errors.New("artist not found")
 	}
 
-	songs, err := s.repo.GetArtistSongs(artistId)
+	var empty_songs *[]models.Song
+
+	songs, err := s.repo.GetArtistSongs(artist.ID)
 	if err != nil {
 		return nil, empty_songs, err
 	}
